FBReader 0.10.0

+++ fbreader/src/formats/docbook/DocBookBookReader.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2004-2008 Geometer Plus <contact@geometerplus.com>
+ * Copyright (C) 2004-2009 Geometer Plus <contact@geometerplus.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -30,7 +30,7 @@ DocBookBookReader::DocBookBookReader(BookModel &model) : BookReader(model) {
 	myReadText = false;
 }
 
-void DocBookBookReader::characterDataHandler(const char *text, int len) {
+void DocBookBookReader::characterDataHandler(const char *text, size_t len) {
 	addDataToBuffer(text, len);
 }
