+++ fbreader/src/formats/html/HtmlDescriptionReader.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2004-2008 Geometer Plus <contact@geometerplus.com>
+ * Copyright (C) 2004-2009 Geometer Plus <contact@geometerplus.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -45,7 +45,7 @@ bool HtmlDescriptionReader::tagHandler(const HtmlTag &tag) {
 	return tag.Name != "BODY";
 }
 
-bool HtmlDescriptionReader::characterDataHandler(const char *text, int len, bool) {
+bool HtmlDescriptionReader::characterDataHandler(const char *text, size_t len, bool) {
 	if (myReadTitle) {
 		myDescription.title().append(text, len);
 	}
