+++ fbreader/src/formats/pdb/MobipocketHtmlBookReader.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2004-2008 Geometer Plus <contact@geometerplus.com>
+ * Copyright (C) 2004-2009 Geometer Plus <contact@geometerplus.com>
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -285,7 +285,7 @@ MobipocketHtmlBookReader::MobipocketHtmlBookReader(const std::string &fileName, 
 	setProcessPreTag(false);
 }
 
-bool MobipocketHtmlBookReader::characterDataHandler(const char *text, int len, bool convert) {
+bool MobipocketHtmlBookReader::characterDataHandler(const char *text, size_t len, bool convert) {
 	myTocReader.appendText(text, len);
 	return HtmlBookReader::characterDataHandler(text, len, convert);
 }
